Moves helper into support folder
Wraps SubscripionHelper module around definitions
